Fixed conversion warning in form builder

+++ src/TGUI/FormBuilder/FormBuilder.cpp
@@ -316,7 +316,7 @@ void FormBuilder::resize(unsigned int width, unsigned int height)
                                          - activeForm->window->getBorders().bottom - activeForm->window->getTitleBarHeight()) / 2.0f);
 
         tgui::ComboBox::Ptr widgetSelector = propertiesWindow->get("WidgetSelector");
-        widgetSelector->setItemsToDisplay((propertiesWindow->getSize().y * 17 / 20) / TGUI_MAXIMUM(10, widgetSelector->getSize().y));
+        widgetSelector->setItemsToDisplay(static_cast<unsigned int>((propertiesWindow->getSize().y * 17 / 20) / TGUI_MAXIMUM(10, widgetSelector->getSize().y)));
     }
 
     window.setView(sf::View(sf::FloatRect(0, 0, static_cast<float>(width), static_cast<float>(height))));
